tx/puller: cycle through missing transactions
The set of missing transaction IDs is now cycled through as transactions
are requested to be pulled from surrouding peers. For every pull attempt
made with a transaction ID, the transaction ID is moved to the head of
the missing transactions LRU cache.

Added getOrPut() and popOrNull() and moveEntryToFront() to lru.HashMap.

Removed the custom hash function and equality function used for the
transaction pusher's LRU cache and use std.hash_map.AutoContext instead.

Ensure that the eventfd is always rearmed should it need to be rearmed
on every event loop iteration.