rheia: implement key/id generation, handshakes, and alternative lru
Keys and ID's are now generated upon instantiation of a Node. Node ID's
comprise of both an Ed25519 public key and a IPv4/IPv6 address-port
pair.

When a node instantiates a new net.Client, the node sends its ID to the
clients' target address in order to perform a handshake. The client is
expected to respond with its own ID. In subsequent commits, the ID
received from the client is to be stored in the nodes' routing table.

An alternative LRU design was implemented as a benchmark against the
current LRU design. The alternative design splits allocations for both
hash map entries and deque nodes into two. The hope was that as the
alternative design requires far less pointer swapping on insert(), it
would benchmark better in comparison to the current design. However, it
turns out that splitting the allocations up caused more pointer
indirections which in turn caused the alternative LRU design to be 30%
slower than the existing design given a large number of entries and any
arbitrary load factor.

A more ergonomic API was introduced for Rheia's LRU implementations.
Upon fetching an entry from a cache, the entry may be moved to the front
or back of the LRU cache using moveToFront() and moveToBack().

The insert() procedure of Rheia's LRU cache implementation was
simplified such that the prepending of new nodes is done in update().