http: port picohttpparser to zig and create http server skeleton
Ported (most of) picohttpparser to Zig to create a HTTP server for Rheia
nodes. The only thing left to be ported is the chunked body decoder.
Added some unit tests as a sanity check to make sure the ported parser
is working as expected.

Wrote a barebones async HTTP server with backpressure support. For the
time being, a 'hello world' response is emitted.

Cleaned up the main entrypoint now that a HTTP server is initialized on
startup. Should any subprocedures of the main entrypoint fail, the
runtime.Context that all subprocedures depend on gets canceled.

Updated the net.Address parsing rules s.t. if there is no colon in an
address string, then the host is assumed to be 127.0.0.1. If there is a
semicolon, then the host is assumed to be unspecified (0.0.0.0).